simple_music_classifier/pytorch_approach/requirements.txt
http://download.pytorch.org/whl/cu75/torch-0.2.0.post3-cp27-cp27mu-manylinux1_x86_64.whl
torchvision
numpy
scipy
tabulate